Seperate ConsoleDisplay from JobRunner
